Split Rabbit auto-config into separate config classes in order to allow Spring Cloud Stream Rabbit Binder to break dependence on RabbitAutoConfiguration.
